configs: making sure we expand any context variables in settings
The important change is adding _load_settings_from_path which will
now call expand_ctx on the WsSettingsHandler. We create a WsContextData
with an empty list of variables so the only thing that contains any
valid data is the default variables in WsContextData. This means that
we can populate the workspace.json with some built in context variables
that is controlled by bakery. This will enable to have a global
workspace.json.