Fix bug where the context variable BKRY_BUILD_VARIANT is not expanded correctly
Calling the expansion in the handler will make the test work but the
actual execution will fail because we will end up with having
BKRY_BUILD_VARIANT wrongly expanded in the artifacts. We don't actually
need it for anything but passing the test the expansion will work
because it will be expanded later on anyway.